# api/src/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CBC Media API",
		"description": "This is the API specification for the CBC Media API",
		"version": "1.0.1"
	},
	"externalDocs": {
		"description": "Old API docs",
		"url": "https://api.countrysidebible.org/"
	},
	"servers": [
		{
			"url": "http://23.236.59.106/",
			"description": "Josh's Server"
		}
	],
	"tags": [
		{
			"name": "media"
		},
		{
			"name": "teacher"
		},
		{
			"name": "series"
		}
	],
	"paths": {
		"/media": {
			"post": {
				"tags": [
					"media"
				],
				"summary": "Post Media",
				"description": "creates a new Media object and returns the MediaCode",
				"requestBody": {
					"$ref": "#/components/requestBodies/CreateMediaRequest"
				},
				"responses": {
					"201": {
						"description": "Success - Media created"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		},
		"/media/{MediaCode}": {
			"get": {
				"tags": [
					"media"
				],
				"summary": "Get Media by MediaCode",
				"description": "returns a Media object with a matching MediaCode",
				"parameters": [
					{
						"name": "MediaCode",
						"in": "path",
						"description": "The MediaCode of the Media to get",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"$ref": "#/components/schemas/MediaCode"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Media"
								}
							}
						}
					},
					"404": {
						"description": "Not Found - Media not found for specified MediaCode"
					}
				}
			},
			"put": {
				"tags": [
					"media"
				],
				"summary": "Modify Media by MediaCode",
				"description": "modifies an existing Media object with the specified MediaCode",
				"parameters": [
					{
						"name": "MediaCode",
						"in": "path",
						"description": "The MediaCode of the Media to modify",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"$ref": "#/components/schemas/MediaCode"
						}
					}
				],
				"requestBody": {
					"$ref": "#/components/requestBodies/UpdateMediaRequest"
				},
				"responses": {
					"204": {
						"description": "Success - Media object modified"
					},
					"404": {
						"description": "Not Found - Media not found for specified MediaCode"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			},
			"delete": {
				"tags": [
					"media"
				],
				"summary": "Delete Media by MediaCode",
				"description": "deletes the Media object with the specified MediaCode",
				"parameters": [
					{
						"name": "MediaCode",
						"in": "path",
						"description": "The MediaCode of the Media to delete",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"$ref": "#/components/schemas/MediaCode"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Success - Media deleted"
					},
					"404": {
						"description": "Not Found - Media not found for specified MediaCode"
					}
				}
			}
		},
		"/teacher": {
			"post": {
				"tags": [
					"teacher"
				],
				"summary": "Post Teacher",
				"description": "creates a new Teacher and returns the ObjectId",
				"requestBody": {
					"$ref": "#/components/requestBodies/CreateTeacherRequest"
				},
				"responses": {
					"201": {
						"description": "Success - Teacher created"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		},
		"/teacher/{ObjectId}": {
			"get": {
				"tags": [
					"teacher"
				],
				"summary": "Get Teacher by ObjectId",
				"description": "returns a Teacher with a matching ObjectId",
				"parameters": [
					{
						"name": "ObjectId",
						"in": "path",
						"description": "The ObjectId of the teacher to get",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"pattern": "[a-fA-F0-9]{24}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Teacher"
								}
							}
						}
					},
					"404": {
						"description": "Not Found - Teacher not found for specified ObjectId"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			},
			"put": {
				"tags": [
					"teacher"
				],
				"summary": "Modify Teacher by ObjectId",
				"description": "modifies an existing Teacher with the specified ObjectId",
				"parameters": [
					{
						"name": "ObjectId",
						"in": "path",
						"description": "The ObjectId of the teacher to modify",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"pattern": "[a-fA-F0-9]{24}"
						}
					}
				],
				"requestBody": {
					"$ref": "#/components/requestBodies/UpdateTeacherRequest"
				},
				"responses": {
					"204": {
						"description": "Success - Teacher modified"
					},
					"404": {
						"description": "Not Found - Teacher not found for specified ObjectId"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			},
			"delete": {
				"tags": [
					"teacher"
				],
				"summary": "Delete Teacher by ObjectId",
				"description": "deletes the Teacher with the specified ObjectId",
				"parameters": [
					{
						"name": "ObjectId",
						"in": "path",
						"description": "The ObjectId of the teacher to delete",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"pattern": "[a-fA-F0-9]{24}"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Success - Teacher deleted"
					},
					"404": {
						"description": "Not Found - Teacher not found for specified ObjectId"
					},
					"400": {
						"description": "Bad Request"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"MediaCode": {
				"type": "string",
				"description": "a globally unique id composed of the dateRecorded (YYYYMMDD), the time (am or pm), the category, and the subCategory",
				"pattern": "\\d{8}(am|pm)-[a-z]{3}(-[a-z]{3})?",
				"minLength": 14,
				"maxLength": 18
			},
			"Media": {
				"type": "object",
				"required": [
					"mediaCode",
					"category",
					"dateRecorded",
					"filesize",
					"mediacode",
					"teacher",
					"text",
					"title"
				],
				"properties": {
					"mediaCode": {
						"$ref": "#/components/schemas/MediaCode"
					},
					"dateRecorded": {
						"type": "string",
						"description": "the date the Media was recorded",
						"format": "date"
					},
					"title": {
						"maxLength": 150,
						"minLength": 1,
						"type": "string",
						"description": "The title of the Media"
					},
					"teacher": {
						"$ref": "#/components/schemas/SimpleTeacher"
					},
					"text": {
						"type": "string",
						"description": "A bible reference used by the media"
					},
					"category": {
						"type": "string",
						"description": "The category of the Media. Must be one of 6 values",
						"enum": [
							"Adult",
							"Students",
							"Family",
							"Baptisms",
							"Conferences",
							"Seminars",
							"Special Services"
						]
					},
					"subCategory": {
						"type": "string",
						"description": "The subcategory of the Media. Must be a child category of the specified category"
					},
					"series": {
						"$ref": "#/components/schemas/SimpleSeries"
					},
					"part": {
						"type": "number",
						"description": "Parts of a multi-session lession where the title is identical",
						"minimum": 1,
						"maximum": 100
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"audio": {
						"$ref": "#/components/schemas/MediaFile"
					},
					"video": {
						"$ref": "#/components/schemas/MediaFile"
					},
					"videmoId": {
						"type": "integer",
						"description": "The ID of the vimeo video's url (such as 12345 in vimeo.com/12345)"
					},
					"youtubeId": {
						"type": "string",
						"description": "The ID of the youtube video's url (such as f8nmJ5 in youtube.com/watch?v=f8nmJ5)"
					},
					"slidesUrl": {
						"type": "string",
						"format": "url",
						"description": "The url of a file containing slides"
					},
					"outlineUrl": {
						"type": "string",
						"format": "url",
						"description": "The url of a file containing the outline"
					},
					"transcriptUrl": {
						"type": "string",
						"format": "url",
						"description": "The url of a file containing the transcript"
					}
				}
			},
			"SimpleTeacher": {
				"type": "object",
				"required": [
					"_id",
					"firstName",
					"lastName",
					"fullName"
				],
				"properties": {
					"_id": {
						"type": "string",
						"description": "An ObjectId string of the Teacher's ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"firstName": {
						"type": "string",
						"description": "The first name. Use for searching, not displaying"
					},
					"lastName": {
						"type": "string",
						"description": "The last name. Use for searching, not displaying"
					},
					"fullName": {
						"type": "string",
						"description": "The full name used primarily for displaying. May somehow get\nused for searching"
					}
				}
			},
			"SimpleSeries": {
				"type": "object",
				"required": [
					"_id",
					"title"
				],
				"properties": {
					"_id": {
						"type": "string",
						"description": "An ObjectId string of the Series' ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"title": {
						"type": "string",
						"minLength": 150
					},
					"image": {
						"type": "string",
						"format": "url",
						"description": "a url to an image used as a thumbnail (any aspect ratio)"
					},
					"imageSquare": {
						"type": "string",
						"format": "url",
						"description": "a url to a square image used as a thumbnail"
					}
				}
			},
			"Teacher": {
				"type": "object",
				"required": [
					"_id",
					"firstName",
					"lastName",
					"fullName"
				],
				"properties": {
					"_id": {
						"type": "string",
						"description": "An ObjectId string used to uniquely identify this Teacher",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"firstName": {
						"type": "string",
						"description": "The teacher's first name"
					},
					"lastName": {
						"type": "string",
						"description": "The teacher's last name"
					},
					"fullName": {
						"type": "string",
						"description": "The teacher's full name, including all titles, prefixes, etc."
					},
					"bio": {
						"type": "string",
						"description": "A biography of the teacher"
					},
					"imageUrl": {
						"type": "string",
						"description": "A url to an image of the teacher",
						"format": "url"
					}
				}
			},
			"MediaFile": {
				"type": "object",
				"required": [
					"duration",
					"fileSize",
					"url"
				],
				"properties": {
					"duration": {
						"type": "integer",
						"description": "the number of seconds long the MediaFile is"
					},
					"fileSize": {
						"type": "integer",
						"description": "the size (in bytes) of the MediaFile"
					},
					"url": {
						"type": "string",
						"format": "url",
						"description": "the url of the MediaFile"
					}
				}
			},
			"CreateMediaRequest": {
				"type": "object",
				"required": [
					"category",
					"dateRecorded",
					"filesize",
					"mediacode",
					"teacherId",
					"text",
					"title"
				],
				"properties": {
					"dateRecorded": {
						"type": "string",
						"description": "the date the Media was recorded",
						"format": "date"
					},
					"title": {
						"maxLength": 150,
						"minLength": 1,
						"type": "string",
						"description": "The title of the Media"
					},
					"teacherId": {
						"type": "string",
						"description": "An ObjectId string of the series' ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"text": {
						"type": "string",
						"description": "A bible reference used by the media."
					},
					"seriesId": {
						"type": "string",
						"description": "An ObjectId string of the series' ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"category": {
						"type": "string",
						"description": "The category of the Media. Must be one of 6 values",
						"enum": [
							"Adult",
							"Students",
							"Family",
							"Baptisms",
							"Conferences",
							"Seminars",
							"Special Services"
						]
					},
					"subCategory": {
						"type": "string",
						"description": "The subcategory of the Media. Must be a child category of the specified category"
					},
					"part": {
						"type": "number",
						"description": "Parts of a multi-session lession where the title is identical",
						"minimum": 1,
						"maximum": 100
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"UpdateMediaRequest": {
				"type": "object",
				"properties": {
					"title": {
						"maxLength": 150,
						"minLength": 1,
						"type": "string",
						"description": "The title of the media"
					},
					"teacherId": {
						"type": "string",
						"description": "An ObjectId string of the teacher's ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					},
					"text": {
						"type": "string",
						"description": "A bible reference used by the media."
					},
					"seriesId": {
						"type": "string",
						"description": "An ObjectId string of the series' ObjectId",
						"pattern": "[a-fA-F0-9]{24}"
					}
				}
			},
			"CreateTeacherRequest": {
				"type": "object",
				"required": [
					"firstName",
					"lastName",
					"fullName"
				],
				"properties": {
					"firstName": {
						"type": "string",
						"minLength": 1,
						"maxLength": 30,
						"description": "The teacher's first name, nothing more"
					},
					"lastName": {
						"minLength": 1,
						"maxLength": 30,
						"type": "string",
						"description": "The teacher's last name, nothing more"
					},
					"fullName": {
						"type": "string",
						"minLength": 3,
						"maxLength": 60,
						"description": "The teacher's full name, including all titles, prefixes, etc."
					},
					"bio": {
						"type": "string",
						"minLength": 1,
						"maxLength": 2000,
						"description": "A biography of the teacher"
					},
					"imageUrl": {
						"type": "string",
						"description": "A url to an image of the teacher",
						"format": "url"
					}
				}
			},
			"UpdateTeacherRequest": {
				"type": "object",
				"properties": {
					"firstName": {
						"type": "string",
						"minLength": 1,
						"maxLength": 30,
						"description": "The teacher's first name, nothing more"
					},
					"lastName": {
						"minLength": 1,
						"maxLength": 30,
						"type": "string",
						"description": "The teacher's last name, nothing more"
					},
					"fullName": {
						"type": "string",
						"minLength": 3,
						"maxLength": 60,
						"description": "The teacher's full name, including all titles, prefixes, etc."
					},
					"bio": {
						"type": "string",
						"minLength": 1,
						"maxLength": 2000,
						"description": "A biography of the teacher"
					},
					"imageUrl": {
						"type": "string",
						"description": "A url to an image of the teacher",
						"format": "url"
					}
				}
			}
		},
		"requestBodies": {
			"CreateMediaRequest": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/CreateMediaRequest"
						}
					}
				}
			},
			"UpdateMediaRequest": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/UpdateMediaRequest"
						}
					}
				}
			},
			"CreateTeacherRequest": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/CreateTeacherRequest"
						}
					}
				}
			},
			"UpdateTeacherRequest": {
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/UpdateTeacherRequest"
						}
					}
				}
			}
		}
	}
}